requirements.txt
flask
google-generativeai
python-dotenv
pydantic
fpdf2
